fix(backend): 报价审批人取名去除冗余 i64 转换
approved_by 已是 i64，user_map.get(&(id as i64)) 的 i64->i64 转换多余，
触发 clippy::unnecessary_cast（-D warnings 阻断）。

+++ backend/src/services/quotation_ops/crud.rs
@@ -261,7 +261,7 @@ impl QuotationService {
             dto.sales_user_name = user_map.get(&dto.sales_user_id).and_then(|n| n.clone());
             dto.approved_by_name = dto
                 .approved_by
-                .and_then(|id| user_map.get(&(id as i64)).and_then(|n| n.clone()));
+                .and_then(|id| user_map.get(&id).and_then(|n| n.clone()));
         }
         Ok(())
     }
